+++ src/carbonado.rs
@@ -1,3 +1,8 @@
+#[cfg(feature = "server")]
+use crate::info;
+#[cfg(feature = "server")]
+use tokio::fs;
+
 use amplify::hex::ToHex;
 use anyhow::{anyhow, Context, Result};
 use bitcoin_30::secp256k1::{PublicKey, SecretKey};
@@ -7,18 +12,17 @@ pub mod constants;
 
 use crate::{carbonado::constants::FORM, constants::CARBONADO_ENDPOINT};
 
+#[cfg(not(feature = "server"))]
 pub async fn store(sk: &str, name: &str, input: &[u8]) -> Result<()> {
     let level = 15;
     let sk = hex::decode(sk)?;
     let secret_key = SecretKey::from_slice(&sk)?;
     let public_key = PublicKey::from_secret_key_global(&secret_key);
     let pk = public_key.serialize();
-
+    let pk_hex = hex::encode(pk);
     let (body, _encode_info) = carbonado::file::encode(&sk, Some(&pk), input, level)?;
-
     let endpoint = CARBONADO_ENDPOINT.read().await.to_string();
     let name = utf8_percent_encode(name, FORM);
-    let pk_hex = hex::encode(pk);
     let url = format!("{endpoint}/{pk_hex}/{name}");
     let client = reqwest::Client::new();
     let response = client
@@ -44,6 +48,23 @@ pub async fn store(sk: &str, name: &str, input: &[u8]) -> Result<()> {
     }
 }
 
+#[cfg(feature = "server")]
+pub async fn store(sk: &str, name: &str, input: &[u8]) -> Result<()> {
+    let level = 15;
+    let sk = hex::decode(sk)?;
+    let secret_key = SecretKey::from_slice(&sk)?;
+    let public_key = PublicKey::from_secret_key_global(&secret_key);
+    let pk = public_key.serialize();
+    let pk_hex = hex::encode(pk);
+    let (body, _encode_info) = carbonado::file::encode(&sk, Some(&pk), input, level)?;
+
+    let filepath = handle_file(&pk_hex, &name, body.len()).await?;
+    #[cfg(feature = "foo")]
+    fs::write(filepath, body).await?;
+    Ok(())
+}
+
+#[cfg(not(feature = "server"))]
 pub async fn retrieve(sk: &str, name: &str) -> Result<Vec<u8>> {
     let sk = hex::decode(sk)?;
     let secret_key = SecretKey::from_slice(&sk)?;
@@ -79,6 +100,43 @@ pub async fn retrieve(sk: &str, name: &str) -> Result<Vec<u8>> {
         let (_header, decoded) = carbonado::file::decode(&sk, &encoded)?;
         Ok(decoded)
     }
+}
+
+#[cfg(feature = "server")]
+pub async fn retrieve(sk: &str, name: &str) -> Result<Vec<u8>> {
+    let sk = hex::decode(sk)?;
+    let secret_key = SecretKey::from_slice(&sk)?;
+    let public_key = PublicKey::from_secret_key_global(&secret_key);
+    let pk = public_key.to_hex();
+
+    let filepath = handle_file(&pk, &name, 0).await?;
+    let bytes = fs::read(filepath).await?;
+
+    if bytes.is_empty() {
+        Ok(Vec::new())
+    } else {
+        let (_header, decoded) = carbonado::file::decode(&sk, &bytes)?;
+        Ok(decoded)
+    }
+}
+
+#[cfg(feature = "server")]
+pub async fn handle_file(pk: &str, name: &str, bytes: usize) -> Result<std::path::PathBuf> {
+    let directory = std::path::Path::new(
+        &std::env::var("CARBONADO_DIR").unwrap_or("/tmp/bitmaskd/carbonado".to_owned()),
+    )
+    .join(pk);
+    let filepath = directory.join(name);
+
+    fs::create_dir_all(directory).await?;
+
+    if bytes == 0 {
+        info!("read {}", filepath.to_string_lossy());
+    } else {
+        info!("write {bytes} bytes to {}", filepath.to_string_lossy());
+    }
+
+    Ok(filepath)
 }
 
 // Utility functions for handling data of different encodings
